src/requirements/base.txt
django
djangorestframework
markdown
django-filter
uwsgi
MySQL-python
mkdocs
docopt
django-q
django-extensions
django-silk
lintjenkins
django-cors-headers
premailer
djmail